ai/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
gunicorn
railtracks
railtracks-cli
python-dotenv
httpx